This improves the HTTP status codes returned by rdxport.cgi
This is an attempt to streamline the usage of HTTP status codes.
Now the only reason for a 401 is wrong user credentials. Errors
due to unsifficient user privileges now return 403. This is also
used for general system restrictions, such as the maximum number
of cuts inside a cart.
Also some internal server errors were reported as 400 which is
now fixed as well.

Signed-off-by: Christian Pointner <equinox@helsinki.at>